requirements.txt
python-binance
numpy
wandb
pandas
matplotlib